requirements.txt
future==1.0.0
iso8601==2.1.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
pyserial==3.5
PyYAML==6.0.2
qasync==0.27.1
